requirements.txt
accelerate==0.28.0
diffusers==0.25.0
huggingface-hub==0.20.3
insightface==0.7.3
joblib==1.3.2
opencv-python==4.8.0.76
onnxruntime==1.17.1
packaging==24.0
pillow==10.2.0
safetensors~=0.4.2
scikit-learn==1.3.2
scipy==1.10.1
torch==2.2.1
tqdm==4.65.0
transformers==4.39.1
numpy==1.24.4
dlib==19.24.2
